Fix metadata service when FIP is used
The NAT configuration used to configure floating IPs conflicts with
the metadata service DNAT rule. The metadata service will return
an error: "404 Not Found The resource could not be found." when a
floating IP is active on the instance and the instance attempts to
contact metadata over the 169.254.169.254 address.

This is because Felix's FIP SNAT rules are natting the instance's
fixed IP to the floating IP before transmitting the packet to the
metadata API service. The metadata API only supports lookups based
on the fixed IP, and it fails to locate metadata for the floating
IP of the instance, so it returns a 404.

This fix will exclude RETURN from the felix FIP SNAT chain when
the metadata service is being contacted without performing any
SNAT conversion on the packet.

This applies primarily when NatSnatDestination is set to "any".